requirements.txt
cffi
pynacl
scrypt
